mini_project/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-recommends==1.0.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
joblib==1.0.1
kombu==5.1.0
MarkupSafe==2.0.1
numpy==1.21.0
oauthlib==3.1.1
pandas==1.3.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
threadpoolctl==2.2.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
